requirements.txt
cachetools==5.3.1
google-auth==2.23.1
google-auth-oauthlib==1.1.0
gspread==5.11.2
oauthlib==3.2.2
prompt-toolkit==1.0.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
regex==2023.8.8
requests-oauthlib==1.3.1
rsa==4.9
urllib3==2.0.5
